docs/pyproject.toml
[project]
name = "nf-modules-website"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.3.0",
    "httpx>=0.27.0",
    "jinja2>=3.1.6",
    "mkdocs>=1.6.1",
    "mkdocs-awesome-pages-plugin>=2.10.1",
    "mkdocs-material>=9.6.16",
    "pyyaml>=6.0.2",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
]
